evolving.py
import torch
import torch.nn as nn
import torch.nn.functional as F
import tkinter as tk

class EvolvingModel(nn.Module):
    def __init__(self, model: nn.Module):
        super(EvolvingModel, self).__init__()
        self.model = nn.ModuleList([model])
        self.root = tk.Tk()
        self.root.title("Evolving Model")
        self.root.geometry("300x200")
        # Create a console to communicate with the evolving model
        self.console = tk.Text(self.root, height=10, width=40)
        self.console.pack()
        # Create a button to evolve the model
        self.evolve_button = tk.Button(self.root, text="Evolve", command=self.evolve)
        self.evolve_button.pack()
    def evolve(self):
        # Add a new layer to the model
        self.model.append(nn.Linear(10, 10))
        # Update the console
        self.console.insert(tk.END, "Model evolved\n")
    def forward(self, x):
        return self.model[0](x)