requirements.txt
torchvision
tk